feat: replace Next.js portfolio with static site.html as new homepage
- Add standalone site.html with dark mode, pull-string theme toggle, SVG footer icons, timeline-only-on-desktop, and responsive mobile layout

- Move old Next.js portfolio to /old route

- Create index.tsx with iframe to serve site.html at / in dev and prod

- Update build script to copy site.html to out/index.html for production

- Refresh favicons and add logo assets
